Fix spaces add room via server
